\name{clean}
\alias{clean}
\alias{\S4method{clean}{BSysProject}}
\title{
  Remove Files Create by this Project
}
\description{
  Calling \code{clean} will remove the files and folders created by the
  project.
}
\usage{
  \S4method{clean}{BSysProject}(.Object)
}
\arguments{
  \item{.Object}{
    \code{.Object} is an object instance of class \code{BSysProject} that describes the code project.
  }
}
\value{
  this method has no return value.
}
\author{
  \packageAuthor{BuildSys}

  Maintainer: \packageMaintainer{BuildSys}
}
\seealso{
  \code{\link{BuildSys-package}}
}
\note{
  see \code{\link{BuildSys-package}} for examples of use.
}

\keyword{debugging}
\keyword{programming}
\keyword{utilities}
